refactor(complete): Move complete_path to long term location
I'd like to remove `ValueHint` eventually
